# corpus/api/openapi.json
{"openapi":"3.1.0","info":{"title":"Curation Studio","version":"0.1.0"},"paths":{"/api/route/{rid}/topoinfo":{"get":{"summary":"Route Topoinfo","description":"Everything the route card's topo section needs: the topos available in\nthis route's area (crag + sibling sectors), flagged with whether this\nroute already has a line on them.","operationId":"route_topoinfo_api_route__rid__topoinfo_get","parameters":[{"name":"rid","in":"path","required":true,"schema":{"type":"integer","title":"Rid"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/topo/{topo_id}":{"get":{"summary":"Topo Detail","operationId":"topo_detail_api_topo__topo_id__get","parameters":[{"name":"topo_id","in":"path","required":true,"schema":{"type":"integer","title":"Topo Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Delete Topo","description":"Remove a bad/typo'd upload: the topo entry, its lines, and the files.","operationId":"delete_topo_api_topo__topo_id__delete","parameters":[{"name":"topo_id","in":"path","required":true,"schema":{"type":"integer","title":"Topo Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"summary":"Patch Topo","operationId":"patch_topo_api_topo__topo_id__put","parameters":[{"name":"topo_id","in":"path","required":true,"schema":{"type":"integer","title":"Topo Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TopoPatch"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/topomedia":{"post":{"summary":"Upload Topo Photo","operationId":"upload_topo_photo_api_topomedia_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_topo_photo_api_topomedia_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/topo/{topo_id}/line/{route_id}":{"put":{"summary":"Put Line","operationId":"put_line_api_topo__topo_id__line__route_id__put","parameters":[{"name":"topo_id","in":"path","required":true,"schema":{"type":"integer","title":"Topo Id"}},{"name":"route_id","in":"path","required":true,"schema":{"type":"integer","title":"Route Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LineBody"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Del Line","operationId":"del_line_api_topo__topo_id__line__route_id__delete","parameters":[{"name":"topo_id","in":"path","required":true,"schema":{"type":"integer","title":"Topo Id"}},{"name":"route_id","in":"path","required":true,"schema":{"type":"integer","title":"Route Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Index","operationId":"index__get","responses":{"200":{"description":"Successful Response","content":{"text/html":{"schema":{"type":"string"}}}}}}},"/api/enums":{"get":{"summary":"Enums","operationId":"enums_api_enums_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/queue":{"get":{"summary":"Queue","operationId":"queue_api_queue_get","parameters":[{"name":"status","in":"query","required":false,"schema":{"type":"string","default":"draft","title":"Status"}},{"name":"q","in":"query","required":false,"schema":{"type":"string","default":"","title":"Q"}},{"name":"crag","in":"query","required":false,"schema":{"type":"string","default":"","title":"Crag"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/map":{"get":{"summary":"Map Points","description":"Every route with its location path and status — feeds the drill-down\naggregates AND the map. Routes without their own coordinates fall back to\nthe nearest ancestor area that has one (approx=true); lat/lon may be null.","operationId":"map_points_api_map_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/route/{rid}":{"get":{"summary":"Route Detail","operationId":"route_detail_api_route__rid__get","parameters":[{"name":"rid","in":"path","required":true,"schema":{"type":"integer","title":"Rid"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"summary":"Patch Route","operationId":"patch_route_api_route__rid__patch","parameters":[{"name":"rid","in":"path","required":true,"schema":{"type":"integer","title":"Rid"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Body"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/route/{rid}/parkings":{"put":{"summary":"Put Parkings","description":"Replace the route's parking spots: [{label, lat, lon}] — multiple pins, ordered.","operationId":"put_parkings_api_route__rid__parkings_put","parameters":[{"name":"rid","in":"path","required":true,"schema":{"type":"integer","title":"Rid"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"array","items":{"type":"object","additionalProperties":true},"title":"Body"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/route/{rid}/pitches":{"put":{"summary":"Put Pitches","operationId":"put_pitches_api_route__rid__pitches_put","parameters":[{"name":"rid","in":"path","required":true,"schema":{"type":"integer","title":"Rid"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"array","items":{"type":"object","additionalProperties":true},"title":"Body"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/route/{rid}/image/{kind}":{"post":{"summary":"Upload Image","description":"Store an image in corpus/uploads/ and point the route's image blob at it.\nExisting keys (alt, attribution…) are preserved; only url changes.","operationId":"upload_image_api_route__rid__image__kind__post","parameters":[{"name":"rid","in":"path","required":true,"schema":{"type":"integer","title":"Rid"}},{"name":"kind","in":"path","required":true,"schema":{"type":"string","title":"Kind"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_image_api_route__rid__image__kind__post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/sources":{"get":{"summary":"Sources","operationId":"sources_api_sources_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/route/{rid}/refs":{"post":{"summary":"Add Ref","description":"Attach a source link (external_ref) to a route — one per source.","operationId":"add_ref_api_route__rid__refs_post","parameters":[{"name":"rid","in":"path","required":true,"schema":{"type":"integer","title":"Rid"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Body"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/route/{rid}/refs/{source_id}":{"delete":{"summary":"Del Ref","operationId":"del_ref_api_route__rid__refs__source_id__delete","parameters":[{"name":"rid","in":"path","required":true,"schema":{"type":"integer","title":"Rid"}},{"name":"source_id","in":"path","required":true,"schema":{"type":"string","title":"Source Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/route/{rid}/references":{"put":{"summary":"Put References","description":"Replace the route's outbound reference links. Text is stored verbatim;\nthe prefix is parsed metadata (mp-field-mapping.md) — never invented.","operationId":"put_references_api_route__rid__references_put","parameters":[{"name":"rid","in":"path","required":true,"schema":{"type":"integer","title":"Rid"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"array","items":{"type":"object","additionalProperties":true},"title":"Body"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/route/{rid}/guidebooks":{"put":{"summary":"Put Guidebooks","description":"Replace the route's guidebook links; guidebook identity is shared across\nroutes and deduped on isbn (else title) — edits here update every copy.","operationId":"put_guidebooks_api_route__rid__guidebooks_put","parameters":[{"name":"rid","in":"path","required":true,"schema":{"type":"integer","title":"Rid"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"array","items":{"type":"object","additionalProperties":true},"title":"Body"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/route/{rid}/publish":{"post":{"summary":"Publish","operationId":"publish_api_route__rid__publish_post","parameters":[{"name":"rid","in":"path","required":true,"schema":{"type":"integer","title":"Rid"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/route/{rid}/status/{new}":{"post":{"summary":"Set Status","operationId":"set_status_api_route__rid__status__new__post","parameters":[{"name":"rid","in":"path","required":true,"schema":{"type":"integer","title":"Rid"}},{"name":"new","in":"path","required":true,"schema":{"type":"string","title":"New"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/route/{rid}/fieldcheck":{"post":{"summary":"Fieldcheck","operationId":"fieldcheck_api_route__rid__fieldcheck_post","parameters":[{"name":"rid","in":"path","required":true,"schema":{"type":"integer","title":"Rid"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/taxonomy":{"get":{"summary":"Taxonomy","operationId":"taxonomy_api_taxonomy_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/taxonomy/{family}":{"post":{"summary":"Taxonomy Add","operationId":"taxonomy_add_api_taxonomy__family__post","parameters":[{"name":"family","in":"path","required":true,"schema":{"type":"string","title":"Family"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Body"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/taxonomy/{family}/{code}":{"patch":{"summary":"Taxonomy Edit","operationId":"taxonomy_edit_api_taxonomy__family___code__patch","parameters":[{"name":"family","in":"path","required":true,"schema":{"type":"string","title":"Family"}},{"name":"code","in":"path","required":true,"schema":{"type":"string","title":"Code"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Body"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Taxonomy Delete","operationId":"taxonomy_delete_api_taxonomy__family___code__delete","parameters":[{"name":"family","in":"path","required":true,"schema":{"type":"string","title":"Family"}},{"name":"code","in":"path","required":true,"schema":{"type":"string","title":"Code"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/export":{"post":{"summary":"Export","description":"Regenerate the committed corpus.json export from the record.","operationId":"export_api_export_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Body_upload_image_api_route__rid__image__kind__post":{"properties":{"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_image_api_route__rid__image__kind__post"},"Body_upload_topo_photo_api_topomedia_post":{"properties":{"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File"},"area_id":{"type":"integer","title":"Area Id"},"credit":{"type":"string","title":"Credit"},"license":{"type":"string","title":"License","default":"owned"},"permission_note":{"type":"string","title":"Permission Note","default":""},"title":{"type":"string","title":"Title","default":""}},"type":"object","required":["file","area_id","credit"],"title":"Body_upload_topo_photo_api_topomedia_post"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LineBody":{"properties":{"line":{"items":{},"type":"array","title":"Line"},"pitches":{"anyOf":[{"items":{},"type":"array"},{"type":"null"}],"title":"Pitches"},"descent":{"anyOf":[{"items":{},"type":"array"},{"type":"null"}],"title":"Descent"}},"type":"object","required":["line"],"title":"LineBody"},"TopoPatch":{"properties":{"title":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Title"},"belay_size":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Belay Size"},"status":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Status"}},"type":"object","title":"TopoPatch"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}